Fix string termination bug
This bug was hard to track down. We only want to terminate string parsing if the quote char we find is the same type that we started the string with. IE if the string started with " it should be terminated only by ", not '. In addition to this, when we find a terminated string, we need to write the buffer into a new token and append it to the token array.